Moved createEntity into EntityFactory
